demo/requirements.txt
# pip install -r requirements.txt
langchain
flask
openai
faiss-cpu
flask-cors
pypdf
tqdm